+++ JavaWeb/src/main/java/ua/karazin/ilyin/javaweb/entity/User.java
@@ -2,7 +2,7 @@ package ua.karazin.ilyin.javaweb.entity;
 
 public class User {
 
-    private long id;
+    private int id;
     private String login;
     private String password;
     private Role role;
@@ -10,18 +10,18 @@ public class User {
     public User() {
     }
 
-    public User(long id, String login, String password, Role role) {
+    public User(int id, String login, String password, Role role) {
         this.id = id;
         this.login = login;
         this.password = password;
         this.role = role;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
